refactor: 금액 BigDecimal에서 long으로 타입 변경

+++ src/main/java/com/kakaopay/scattering/domain/MoneyDivideStrategy.java
@@ -1,8 +1,6 @@
 package com.kakaopay.scattering.domain;
 
-import java.math.BigDecimal;
-
 @FunctionalInterface
 public interface MoneyDivideStrategy {
-    ScatteredMonies divide(BigDecimal money, int count);
+    ScatteredMonies divide(long money, int count);
 }
